coffee_pos/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
pandas==2.0.3
psycopg2-binary==2.9.7
gunicorn==21.2.0
